+++ frontend/src/App.js
@@ -2,7 +2,10 @@ import React, { useEffect, useState } from "react";
 import "./App.css";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import { initializeApp } from 'firebase/app';
-import { getAuth, onAuthStateChanged, signInWithEmailAndPassword, createUserWithEmailAndPassword, signOut } from 'firebase/auth';
+import { getAuth, onAuthStateChanged } from 'firebase/auth';
+import Login from './components/Login';
+import Dashboard from './components/Dashboard';
+import Layout from './components/Layout';
 import axios from "axios";
 
 // Firebase configuration - Replace with your actual config
@@ -23,7 +26,7 @@ const BACKEND_URL = process.env.REACT_APP_BACKEND_URL;
 const API = `${BACKEND_URL}/api`;
 
 // Auth Context
-const AuthContext = React.createContext();
+export const AuthContext = React.createContext();
 
 // Auth Provider Component
 const AuthProvider = ({ children }) => {
@@ -69,380 +72,6 @@ const AuthProvider = ({ children }) => {
     <AuthContext.Provider value={{ user, loading, auth }}>
       {children}
     </AuthContext.Provider>
-  );
-};
-
-// Login Component
-const Login = () => {
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [isRegistering, setIsRegistering] = useState(false);
-  const [error, setError] = useState('');
-
-  const handleAuth = async (e) => {
-    e.preventDefault();
-    setError('');
-    
-    try {
-      if (isRegistering) {
-        await createUserWithEmailAndPassword(auth, email, password);
-      } else {
-        await signInWithEmailAndPassword(auth, email, password);
-      }
-    } catch (error) {
-      setError(error.message);
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-100 flex items-center justify-center">
-      <div className="max-w-md w-full bg-white rounded-lg shadow-md p-8">
-        <h2 className="text-2xl font-bold text-center mb-6">
-          {isRegistering ? 'Create Account' : 'Sign In'}
-        </h2>
-        
-        {error && (
-          <div className="bg-red-100 border border-red-400 text-red-700 px-4 py-3 rounded mb-4">
-            {error}
-          </div>
-        )}
-        
-        <form onSubmit={handleAuth}>
-          <div className="mb-4">
-            <label className="block text-gray-700 text-sm font-bold mb-2">
-              Email
-            </label>
-            <input
-              type="email"
-              className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:border-blue-500"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              required
-            />
-          </div>
-          
-          <div className="mb-6">
-            <label className="block text-gray-700 text-sm font-bold mb-2">
-              Password
-            </label>
-            <input
-              type="password"
-              className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:border-blue-500"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              required
-            />
-          </div>
-          
-          <button
-            type="submit"
-            className="w-full bg-blue-500 text-white py-2 px-4 rounded-lg hover:bg-blue-600 transition-colors"
-          >
-            {isRegistering ? 'Create Account' : 'Sign In'}
-          </button>
-        </form>
-        
-        <p className="text-center mt-4 text-gray-600">
-          {isRegistering ? 'Already have an account?' : "Don't have an account?"}
-          <button
-            className="text-blue-500 hover:text-blue-700 ml-1"
-            onClick={() => setIsRegistering(!isRegistering)}
-          >
-            {isRegistering ? 'Sign In' : 'Create Account'}
-          </button>
-        </p>
-      </div>
-    </div>
-  );
-};
-
-// Dashboard Component
-const Dashboard = () => {
-  const [stats, setStats] = useState({});
-  const [loading, setLoading] = useState(true);
-  const { user } = React.useContext(AuthContext);
-
-  useEffect(() => {
-    const fetchStats = async () => {
-      try {
-        const response = await axios.get(`${API}/dashboard/stats`);
-        setStats(response.data);
-      } catch (error) {
-        console.error('Error fetching stats:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchStats();
-  }, []);
-
-  if (loading) {
-    return (
-      <div className="flex justify-center items-center h-64">
-        <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-500"></div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="space-y-6">
-      <div className="flex justify-between items-center">
-        <h1 className="text-3xl font-bold text-gray-900">Dashboard</h1>
-        <div className="text-sm text-gray-600">
-          Welcome back, {user?.name}!
-        </div>
-      </div>
-
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
-        <div className="bg-white p-6 rounded-lg shadow-md">
-          <h3 className="text-lg font-semibold text-gray-700 mb-2">Total Pages</h3>
-          <p className="text-3xl font-bold text-blue-600">{stats.total_pages || 0}</p>
-        </div>
-        
-        <div className="bg-white p-6 rounded-lg shadow-md">
-          <h3 className="text-lg font-semibold text-gray-700 mb-2">Total Articles</h3>
-          <p className="text-3xl font-bold text-green-600">{stats.total_articles || 0}</p>
-        </div>
-        
-        <div className="bg-white p-6 rounded-lg shadow-md">
-          <h3 className="text-lg font-semibold text-gray-700 mb-2">Published</h3>
-          <p className="text-3xl font-bold text-purple-600">
-            {(stats.published_pages || 0) + (stats.published_articles || 0)}
-          </p>
-        </div>
-        
-        <div className="bg-white p-6 rounded-lg shadow-md">
-          <h3 className="text-lg font-semibold text-gray-700 mb-2">Drafts</h3>
-          <p className="text-3xl font-bold text-orange-600">
-            {(stats.draft_pages || 0) + (stats.draft_articles || 0)}
-          </p>
-        </div>
-      </div>
-
-      <div className="bg-white p-6 rounded-lg shadow-md">
-        <h2 className="text-xl font-bold text-gray-900 mb-4">MCP Tool Testing</h2>
-        <MCPToolTester />
-      </div>
-    </div>
-  );
-};
-
-// MCP Tool Tester Component
-const MCPToolTester = () => {
-  const [toolName, setToolName] = useState('createPage');
-  const [toolArgs, setToolArgs] = useState('{"title": "Test Page", "content": "This is a test page"}');
-  const [result, setResult] = useState(null);
-  const [loading, setLoading] = useState(false);
-
-  const testTool = async () => {
-    setLoading(true);
-    try {
-      const args = JSON.parse(toolArgs);
-      const response = await axios.post(`${API}/mcp/dispatch`, {
-        tool: toolName,
-        args: args
-      });
-      setResult(response.data);
-    } catch (error) {
-      setResult({ success: false, error: error.message });
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="space-y-4">
-      <div>
-        <label className="block text-sm font-medium text-gray-700 mb-2">
-          Tool Name
-        </label>
-        <select 
-          className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:border-blue-500"
-          value={toolName}
-          onChange={(e) => setToolName(e.target.value)}
-        >
-          <option value="createPage">Create Page</option>
-          <option value="createArticle">Create Article</option>
-          <option value="updatePage">Update Page</option>
-          <option value="createUser">Create User</option>
-        </select>
-      </div>
-
-      <div>
-        <label className="block text-sm font-medium text-gray-700 mb-2">
-          Tool Arguments (JSON)
-        </label>
-        <textarea
-          className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:outline-none focus:border-blue-500"
-          rows="4"
-          value={toolArgs}
-          onChange={(e) => setToolArgs(e.target.value)}
-          placeholder='{"title": "Test Page", "content": "This is a test page"}'
-        />
-      </div>
-
-      <button
-        onClick={testTool}
-        disabled={loading}
-        className="px-4 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-600 disabled:opacity-50"
-      >
-        {loading ? 'Testing...' : 'Test Tool'}
-      </button>
-
-      {result && (
-        <div className="mt-4 p-4 bg-gray-50 rounded-lg">
-          <h4 className="font-semibold mb-2">Result:</h4>
-          <pre className="text-sm overflow-x-auto">
-            {JSON.stringify(result, null, 2)}
-          </pre>
-        </div>
-      )}
-    </div>
-  );
-};
-
-// Pages List Component
-const PagesList = () => {
-  const [pages, setPages] = useState([]);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const fetchPages = async () => {
-      try {
-        const response = await axios.get(`${API}/pages`);
-        setPages(response.data);
-      } catch (error) {
-        console.error('Error fetching pages:', error);
-      } finally {
-        setLoading(false);
-      }
-    };
-
-    fetchPages();
-  }, []);
-
-  if (loading) {
-    return (
-      <div className="flex justify-center items-center h-64">
-        <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-500"></div>
-      </div>
-    );
-  }
-
-  return (
-    <div className="space-y-6">
-      <div className="flex justify-between items-center">
-        <h1 className="text-3xl font-bold text-gray-900">Pages</h1>
-      </div>
-
-      <div className="bg-white shadow-md rounded-lg overflow-hidden">
-        <table className="min-w-full divide-y divide-gray-200">
-          <thead className="bg-gray-50">
-            <tr>
-              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                Title
-              </th>
-              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                Status
-              </th>
-              <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">
-                Created
-              </th>
-            </tr>
-          </thead>
-          <tbody className="bg-white divide-y divide-gray-200">
-            {pages.map((page) => (
-              <tr key={page.id}>
-                <td className="px-6 py-4 whitespace-nowrap">
-                  <div className="text-sm font-medium text-gray-900">{page.title}</div>
-                  <div className="text-sm text-gray-500">{page.slug}</div>
-                </td>
-                <td className="px-6 py-4 whitespace-nowrap">
-                  <span className={`px-2 py-1 inline-flex text-xs leading-5 font-semibold rounded-full ${
-                    page.status === 'published' ? 'bg-green-100 text-green-800' : 'bg-yellow-100 text-yellow-800'
-                  }`}>
-                    {page.status}
-                  </span>
-                </td>
-                <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-500">
-                  {new Date(page.created_at).toLocaleDateString()}
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  );
-};
-
-// Main Layout Component
-const Layout = ({ children }) => {
-  const { user, auth } = React.useContext(AuthContext);
-  const [activeTab, setActiveTab] = useState('dashboard');
-
-  const handleSignOut = async () => {
-    try {
-      await signOut(auth);
-    } catch (error) {
-      console.error('Error signing out:', error);
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-100">
-      <nav className="bg-white shadow-lg">
-        <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-          <div className="flex justify-between h-16">
-            <div className="flex items-center">
-              <h1 className="text-xl font-bold text-gray-900">MCP-CMS</h1>
-            </div>
-            
-            <div className="flex items-center space-x-4">
-              <button
-                onClick={() => setActiveTab('dashboard')}
-                className={`px-3 py-2 rounded-md text-sm font-medium ${
-                  activeTab === 'dashboard' 
-                    ? 'bg-blue-500 text-white' 
-                    : 'text-gray-700 hover:text-gray-900'
-                }`}
-              >
-                Dashboard
-              </button>
-              
-              <button
-                onClick={() => setActiveTab('pages')}
-                className={`px-3 py-2 rounded-md text-sm font-medium ${
-                  activeTab === 'pages' 
-                    ? 'bg-blue-500 text-white' 
-                    : 'text-gray-700 hover:text-gray-900'
-                }`}
-              >
-                Pages
-              </button>
-              
-              <div className="flex items-center space-x-2">
-                <span className="text-sm text-gray-700">{user?.name}</span>
-                <button
-                  onClick={handleSignOut}
-                  className="text-sm text-red-600 hover:text-red-800"
-                >
-                  Sign Out
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </nav>
-
-      <main className="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
-        <div className="px-4 py-6 sm:px-0">
-          {activeTab === 'dashboard' && <Dashboard />}
-          {activeTab === 'pages' && <PagesList />}
-        </div>
-      </main>
-    </div>
   );
 };
