+++ frontend/src/components/Dashboard.js
@@ -1,0 +1,78 @@
+import React, { useState, useEffect, useContext } from 'react';
+import axios from 'axios';
+import { AuthContext } from '../App';
+import MCPToolTester from './MCPToolTester';
+
+const BACKEND_URL = process.env.REACT_APP_BACKEND_URL;
+const API = `${BACKEND_URL}/api`;
+
+const Dashboard = () => {
+  const [stats, setStats] = useState({});
+  const [loading, setLoading] = useState(true);
+  const { user } = useContext(AuthContext);
+
+  useEffect(() => {
+    const fetchStats = async () => {
+      try {
+        const response = await axios.get(`${API}/dashboard/stats`);
+        setStats(response.data);
+      } catch (error) {
+        console.error('Error fetching stats:', error);
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchStats();
+  }, []);
+
+  if (loading) {
+    return (
+      <div className="flex justify-center items-center h-64">
+        <div className="animate-spin rounded-full h-12 w-12 border-b-2 border-blue-500"></div>
+      </div>
+    );
+  }
+
+  return (
+    <div className="space-y-6">
+      <div className="flex justify-between items-center">
+        <h1 className="text-3xl font-bold text-gray-900">Dashboard</h1>
+        <div className="text-sm text-gray-600">Welcome back, {user?.name}!</div>
+      </div>
+
+      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-4 gap-6">
+        <div className="bg-white p-6 rounded-lg shadow-md">
+          <h3 className="text-lg font-semibold text-gray-700 mb-2">Total Pages</h3>
+          <p className="text-3xl font-bold text-blue-600">{stats.total_pages || 0}</p>
+        </div>
+
+        <div className="bg-white p-6 rounded-lg shadow-md">
+          <h3 className="text-lg font-semibold text-gray-700 mb-2">Total Articles</h3>
+          <p className="text-3xl font-bold text-green-600">{stats.total_articles || 0}</p>
+        </div>
+
+        <div className="bg-white p-6 rounded-lg shadow-md">
+          <h3 className="text-lg font-semibold text-gray-700 mb-2">Published</h3>
+          <p className="text-3xl font-bold text-purple-600">
+            {(stats.published_pages || 0) + (stats.published_articles || 0)}
+          </p>
+        </div>
+
+        <div className="bg-white p-6 rounded-lg shadow-md">
+          <h3 className="text-lg font-semibold text-gray-700 mb-2">Drafts</h3>
+          <p className="text-3xl font-bold text-orange-600">
+            {(stats.draft_pages || 0) + (stats.draft_articles || 0)}
+          </p>
+        </div>
+      </div>
+
+      <div className="bg-white p-6 rounded-lg shadow-md">
+        <h2 className="text-xl font-bold text-gray-900 mb-4">MCP Tool Testing</h2>
+        <MCPToolTester />
+      </div>
+    </div>
+  );
+};
+
+export default Dashboard;
